Updated site with all new leaflet plugins and goodies
1. search - same as before
2. location - I tried it once but didn't found it useful, using mobile
with this site it is very useful, decided to add it back.
3. load gpx, kml and probably another format to the map using the folder
icon - let me know what you think about the colors if it is visible
enough.
4. added Google imagery and overlay - currently the overlay and the
regular map are creating some duplication - we need to think if we want
to split the maps
I can manage to work around this with javascript code if needed i
guess...

still missing:
1. info
2. legend - I want to move the map to the Kinneret and add the ability
for the user to click "back", it will reduce maintenance...
3. edit button - still need to understand the whole map control of
leaflet...